fix the third case
