+++ packages/cli/src/commands/__tests__/deploy.test.ts
@@ -1,33 +1,27 @@
 import { afterEach, expect, vi } from 'vitest'
-import { decidePackageName, deploy } from '../deploy'
+import { deploy } from '../deploy'
+import { decidePackageName, selectApiScope } from '../deploy/helper'
 import { getPersonalAccessToken } from '../../utils'
 
 vi.mock('../../utils')
-
-vi.mock('../deploy', async () => {
-  const actualDeploy = await vi.importActual<typeof import('../deploy')>(
-    '../deploy',
-  )
-
-  return {
-    ...actualDeploy,
-    decidePackageName: vi.fn(),
-  }
-})
+vi.mock('../deploy/helper')
 
 describe('deploy', () => {
+  const exit = vi.fn(() => {
+    // eslint-disable-next-line functional/no-throw-statement
+    throw new Error('Exiting Process Error')
+  })
+
+  beforeEach(() => {
+    vi.clearAllMocks()
+    vi.stubGlobal('process', { exit })
+  })
+
   afterEach(() => {
-    vi.resetAllMocks()
     vi.unstubAllGlobals()
   })
 
   it('should exit when getPersonalAccessToken returns an error', async () => {
-    const exit = vi.fn(() => {
-      // eslint-disable-next-line functional/no-throw-statement
-      throw new Error('Exiting Process Error')
-    })
-    vi.stubGlobal('process', { exit })
-
     vi.mocked(getPersonalAccessToken).mockImplementation((params) => {
       return Promise.resolve({ error: true, message: 'Test Error' })
     })
@@ -39,12 +33,6 @@ describe('deploy', () => {
   })
 
   it('should exit when decidePackageName returns invalid name', async () => {
-    const exit = vi.fn(() => {
-      // eslint-disable-next-line functional/no-throw-statement
-      throw new Error('Exiting Process Error')
-    })
-    vi.stubGlobal('process', { exit: exit })
-
     vi.mocked(getPersonalAccessToken).mockImplementation((params) => {
       return Promise.resolve({ error: false, token: 'token' })
     })
@@ -59,13 +47,7 @@ describe('deploy', () => {
     expect(exit).toHaveBeenCalledTimes(1)
   })
 
-  it.only('should exit when scope is not provided', async () => {
-    const exit = vi.fn(() => {
-      // eslint-disable-next-line functional/no-throw-statement
-      throw new Error('Exiting Process Error')
-    })
-    vi.stubGlobal('process', { exit: exit })
-
+  it('should exit when scope is not provided', async () => {
     vi.mocked(getPersonalAccessToken).mockImplementation((params) => {
       return Promise.resolve({ error: false, token: 'token' })
     })
@@ -79,6 +61,9 @@ describe('deploy', () => {
     )
 
     expect(exit).toHaveBeenCalledTimes(1)
+    expect(getPersonalAccessToken).toHaveBeenCalledOnce()
+    expect(decidePackageName).toHaveBeenCalledOnce()
+    expect(selectApiScope).toHaveBeenCalledTimes(0)
   })
 })
